+++ src/main/kotlin/com/inno/soramitsu/insurance/RESTserver/dao/InsuranceRepository.kt
@@ -9,7 +9,6 @@ import org.springframework.data.jpa.repository.Modifying
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
 import org.springframework.stereotype.Repository
-import java.math.BigInteger
 import javax.transaction.Transactional
 
 /**
@@ -28,22 +27,22 @@ interface InsuranceRepository : JpaRepository<Insurance, Long> {
     fun insertNewInsuranceRequest(@Param("newRequest") newRequest: Insurance)
 
     @Transactional
-    fun findByCompanyCompanyid(@Param("companyId") companyId: BigInteger): List<Insurance>
+    fun findByCompanyCompanyid(@Param("companyId") companyId: Long): List<Insurance>
 
     @Transactional
-    fun findByCompanyCompanyidAndStatus(@Param("companyId") companyId: BigInteger,
+    fun findByCompanyCompanyidAndStatus(@Param("companyId") companyId: Long,
                                          @Param("status") status: String): List<Insurance>
 
     @Transactional
     @Modifying
     @Query(value = "update insurance_request set status = :#{#status} where insurance_request_id = :#{#insuranceId}", nativeQuery = true)
-    fun updateInsuranceStatus(@Param("insuranceId") insuranceId: BigInteger, @Param("status") status: String)
+    fun updateInsuranceStatus(@Param("insuranceId") insuranceId: Long, @Param("status") status: String)
 
     @Transactional
     fun findByUserEmail(@Param("email") email: String): List<Insurance>
 
     @Transactional
     @Query(value = "SELECT * FROM insurance_request ir WHERE ir.insurance_request_id=:#{#id} ", nativeQuery = true)
-    fun findByInsuranceRequestId(@Param("id") id: BigInteger): Insurance
+    fun findByInsuranceRequestId(@Param("id") id: Long): Insurance
 
 }
